% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{train_models_summary}
\alias{train_models_summary}
\title{Train Models and Summarize Results}
\usage{
train_models_summary(data, independent_var, dependent_vars)
}
\arguments{
\item{data}{A dataframe containing the independent and dependent variables.}

\item{independent_var}{A string specifying the name of the independent variable.}

\item{dependent_vars}{A vector of strings specifying the names of dependent variables.}
}
\value{
A dataframe summarizing RMSE and R-squared for each model (KSVM, SVM, Linear Regression and Random Forest) and dependent variable.
}
\description{
Train Models and Summarize Results
}
\examples{
train_models_summary(data = read.csv("Electricity1970.csv"), independent_var = "cost", dependent_vars = c("labor", "capital", "fuel"))

data <- read.csv("Electricity1970.csv")
train_models_summary(data, independent_var = "cost", dependent_vars = c("labor", "capital", "fuel"))




}
